+++ internal/provider/model/task.go
@@ -25,7 +25,7 @@ import (
 // ----------------------------------------
 type taskFrequency struct {
 	Schedule       types.String  `tfsdk:"schedule"`
-	StartDate      types.Int32   `tfsdk:"start_date"`
+	StartDate      types.Int64   `tfsdk:"start_date"`
 	TimezoneOffset types.String  `tfsdk:"timezone_offset"`
 	RecurringDays  []types.Int32 `tfsdk:"recurring_days"`
 	CronExpression types.String  `tfsdk:"cron_expression"`
@@ -38,12 +38,7 @@ func (f *taskFrequency) ToApiModel(api *v3.FrequencyXO) {
 		api.RecurringDays = append(api.RecurringDays, rd.ValueInt32())
 	}
 	api.Schedule = f.Schedule.ValueString()
-	if f.StartDate.ValueInt32Pointer() != nil {
-		val := int64(*f.StartDate.ValueInt32Pointer())
-		api.StartDate = &val
-	} else {
-		api.StartDate = nil
-	}
+	api.StartDate = f.StartDate.ValueInt64Pointer()
 	api.TimeZoneOffset = f.TimezoneOffset.ValueStringPointer()
 }
 
@@ -86,16 +81,19 @@ type BaseTaskModel struct {
 
 func (m *BaseTaskModel) MapFromApi(api *v3.TaskXO) {
 	m.Id = types.StringPointerValue(api.Id)
-	m.Name = types.StringPointerValue(api.Name)
-	// m.Type = types.StringPointerValue(api.Type)
+	if api.Name != nil {
+		m.Name = types.StringPointerValue(api.Name)
+	}
 }
 
 func (m *BaseTaskModel) toApiCreateModel() *v3.TaskTemplateXO {
 	api := v3.NewTaskTemplateXOWithDefaults()
 	api.Name = m.Name.ValueString()
 	api.Enabled = m.Enabled.ValueBool()
-	api.Frequency = *v3.NewFrequencyXO(m.Frequency.Schedule.ValueString())
-	m.Frequency.ToApiModel(&api.Frequency)
+	if m.Frequency != nil {
+		api.Frequency = *v3.NewFrequencyXO(m.Frequency.Schedule.ValueString())
+		m.Frequency.ToApiModel(&api.Frequency)
+	}
 	api.AlertEmail = m.AlertEmail.ValueStringPointer()
 	api.NotificationCondition = m.NotificationCondition.ValueString()
 	return api
@@ -105,8 +103,10 @@ func (m *BaseTaskModel) toApiUpdateModel() *v3.UpdateTaskRequest {
 	api := v3.NewUpdateTaskRequestWithDefaults()
 	api.Name = m.Name.ValueString()
 	api.Enabled = m.Enabled.ValueBool()
-	api.Frequency = *v3.NewFrequencyXO(m.Frequency.Schedule.ValueString())
-	m.Frequency.ToApiModel(&api.Frequency)
+	if m.Frequency != nil {
+		api.Frequency = *v3.NewFrequencyXO(m.Frequency.Schedule.ValueString())
+		m.Frequency.ToApiModel(&api.Frequency)
+	}
 	api.AlertEmail = m.AlertEmail.ValueStringPointer()
 	api.NotificationCondition = m.NotificationCondition.ValueString()
 	return api